Add jdbc dao's

+++ homework-05/src/main/java/ru/otus/homework05/dao/BookDao.java
@@ -4,8 +4,9 @@ import ru.otus.homework05.domain.Book;
 
 import java.util.List;
 
+
 public interface BookDao {
-    int count();
+    long count();
 
     void insert(Book book);
